requirements.txt
telethon
python-decouple
python-dotenv
akinator
apscheduler
aiohttp
bs4
enhancer>=0.3.4
git+https://github.com/New-dev0/instagrapi.git@39df1b1#egg=instagrapi
git+https://github.com/buddhhu/img2html.git@c44170d#egg=img2html
heroku3
gitpython
google-api-python-client
htmlwebshot
lottie
lxml
numpy>=1.21.2
oauth2client
opencv-python-headless
pillow>=7.0.0
profanitydetector
psutil
redis 
pypdf2>=1.26.0
pytz
qrcode
requests
tabulate
telegraph
tgcrypto
